refactor(svelte): split surface-state tests and extract a11y descriptions
Move createSurfaceState composite tests into a shared harness and three
responsibility-based files (construction/effects, brush, input lifecycle).
Extract buildSurfaceDescription for pure unit coverage while preserving
lazy inspectConfig reads for Svelte dependency tracking. Characterize
select-end commit-before-emit ordering.

Closes #396
